some changes to get dtn working on raven, contiki-mac crashes the raven
(was: some changes to get dtn working on raven, contiki-mac crashes the
raven)

+++ core/net/dtn/sdnv.h
@@ -3,7 +3,7 @@
 #include <stdlib.h>
 typedef uint8_t *sdnv_t;
 
-int sdnv_encode(uint64_t val, uint8_t* bp, size_t len);
-size_t sdnv_encoding_len(uint64_t val);
-int sdnv_decode(const uint8_t* bp, size_t len, uint64_t* val);
+int sdnv_encode(uint32_t val, uint8_t* bp, size_t len);
+size_t sdnv_encoding_len(uint32_t val);
+int sdnv_decode(const uint8_t* bp, size_t len, uint32_t* val);
 size_t sdnv_len(const uint8_t* bp);
